ui: fix six low-severity Qt UI bugs and harden marker listener
- TimeRuler: derive the visible time window from the painter clip rect
  and iterate tick/tempo loops only over it, instead of repainting the
  full 300s bounding rect every frame (significant redraw cost at high
  zoom).
- AudioWaveformWidget: accept() the wheel event in both branches so
  horizontal-scroll doesn't also bubble up to an ancestor scroll area.
- MainWindow: cast tabButtons.size() to int in three comparisons to
  silence -Wsign-compare (matches the ModulationWidget pattern).
- PhraseGeneratorDialog: use the 4-arg connect form for the note-range
  spinboxes so Qt auto-disconnects if the dialog is destroyed
  (per AGENTS.md signal/slot rule).
- MixerStripWidget: add an in-declaration `= -1` initializer to
  trackIndex as a defensive guard against future ctor refactors.
- TimelineView: filter marker child-add/remove by parentTree type
  (MARKER_LIST) rather than comparing against the stored markerListTree
  member, which can go stale on swap paths that bypass the root-add
  branch (per AGENTS.md "filter by type" guidance).

+++ src/ui/MainWindow.cpp
@@ -459,11 +459,11 @@ void MainWindow::setupBottomPanel()
     });
 
     connect(bottomStack, &QStackedWidget::currentChanged, this, [this](int index) {
-        if (index < 0 || index >= tabButtons.size()) return;
-        for (int i = 0; i < tabButtons.size(); ++i)
+        if (index < 0 || index >= static_cast<int>(tabButtons.size())) return;
+        for (int i = 0; i < static_cast<int>(tabButtons.size()); ++i)
             tabButtons[i]->setChecked(i == index);
     });
-    if (bottomStack->currentIndex() >= 0 && bottomStack->currentIndex() < tabButtons.size())
+    if (bottomStack->currentIndex() >= 0 && bottomStack->currentIndex() < static_cast<int>(tabButtons.size()))
         tabButtons[bottomStack->currentIndex()]->setChecked(true);
 
     bottomStack->setCurrentIndex(0);
